requirements.txt
click==7.0
mlflow==1.14.1
keras==2.3.1
tensorflow==2.1.0
docker==4.1.0
jsonschema==3.0.2
matplotlib==3.1.1
numpy==1.19.5
pandas==0.25.3
pandas-profiling==2.5.0
PyYAML==5.3
scikit-learn==0.21.3
seaborn==0.9.0
six==1.15.0
oyaml==0.9
plotly==4.5.1
fastapi==0.63.0
uvicorn==0.12.3

kubernetes
pickle
couler
seldon-core=1.7.0
boto3
